requirements.txt
numpy
opencv-python
pillow
matplotlib
scikit-learn
scipy
tensorflow
tqdm
mediapipe
sklearn.preprocessing